ref: moved render_template, current_user_optional and templates to a config files
