+++ src/auth/manager.py
@@ -1,5 +1,5 @@
 import uuid
-from typing import Optional, Any
+from typing import Any, Optional
 
 from fastapi import Depends
 from fastapi.requests import Request
