+++ src/frontend/evaluations.py
@@ -2,7 +2,8 @@ from fastapi import APIRouter, Depends
 from fastapi.requests import Request
 from fastapi.responses import HTMLResponse, RedirectResponse
 
-from src.frontend.root import current_user_optional, render_template
+from src.auth.auth import current_user_optional
+from src.frontend.config import render_template
 from src.models import User
 
 router = APIRouter()
