+++ src/frontend/tasks.py
@@ -2,8 +2,9 @@ from fastapi import APIRouter, Depends
 from fastapi.requests import Request
 from fastapi.responses import HTMLResponse, RedirectResponse
 
+from src.auth.auth import current_user_optional
 from src.dependencies.deps import SessionDep
-from src.frontend.root import current_user_optional, render_template
+from src.frontend.config import render_template
 from src.models import User
 from src.services.task_service import list_tasks
 from src.services.user_service import get_all_users
